LLM/filter_message/requirements.txt
openai
pyyaml
pymysql
black